requirements.txt
# Function dependencies, for example:
# package>=version
tensorflow==2.3.0
google-cloud-storage==1.38.0
scikit-learn==0.22.2
google-cloud-bigquery==2.17.0